# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gem",
        "description": "gem is project about transportation",
        "version": "1"
    },
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "vehicles"
                ],
                "summary": "get all vehicles",
                "description": "get all vehicles",
                "operationId": "043b3ce51f006f946511103cfe1c2ca1",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "vehicles"
                ],
                "summary": "create vehicle",
                "description": "create vehicle",
                "operationId": "74e9ae3e5f1ee2f69f3a177820d204b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "number_plate": {
                                        "type": "string"
                                    },
                                    "number_plate_zone": {
                                        "type": "string"
                                    },
                                    "number_plate_type": {
                                        "type": "integer"
                                    },
                                    "numbering_location_code": {
                                        "type": "integer"
                                    },
                                    "numbering_location": {
                                        "type": "string"
                                    },
                                    "VIN": {
                                        "type": "string"
                                    },
                                    "allowed_driver_licence": {
                                        "type": "integer"
                                    },
                                    "vehicle_type": {
                                        "type": "string"
                                    },
                                    "owner_mobile": {
                                        "type": "string"
                                    },
                                    "production": {
                                        "type": "integer"
                                    },
                                    "loading_type": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "inspection_number": {
                                        "type": "string"
                                    },
                                    "inspection_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "insurance_number": {
                                        "type": "string"
                                    },
                                    "insurance_expire": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "brand_code": {
                                        "type": "integer"
                                    },
                                    "brand_title": {
                                        "type": "string"
                                    },
                                    "number_of_allowed_draft": {
                                        "type": "integer"
                                    },
                                    "day_of_allowed_draft": {
                                        "type": "integer"
                                    },
                                    "number_of_allowed_bill_of_lading": {
                                        "type": "integer"
                                    },
                                    "day_of_allowed_bill_of_lading": {
                                        "type": "integer"
                                    },
                                    "number_of_unconfirmed_draft": {
                                        "type": "integer"
                                    },
                                    "number_of_unconfirmed_bill_of_lading": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "weight": {
                                        "type": "integer"
                                    },
                                    "loading_length": {
                                        "type": "integer"
                                    },
                                    "loading_width": {
                                        "type": "integer"
                                    },
                                    "has_roll": {
                                        "type": "integer"
                                    },
                                    "chassis_number": {
                                        "type": "string"
                                    },
                                    "document_number": {
                                        "type": "string"
                                    },
                                    "violation": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "engine_number": {
                                        "type": "string"
                                    },
                                    "document_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "owner_status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/api/vehicles/load-navi-by-smart-card-no": {
            "post": {
                "tags": [
                    "vehicles"
                ],
                "summary": "loading vehicle by smart card number ",
                "description": "loading vehicle by smart card number ",
                "operationId": "8070ec1d62b5acdd1ba9f0a69ce5d9e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "card_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/vehicles/load-navi-by-smart-card-no-full": {
            "post": {
                "tags": [
                    "vehicles"
                ],
                "summary": "loading vehicle by smart card number full",
                "description": "loading vehicle by smart card number full",
                "operationId": "e2a9105b98ffee20fe79cb1e0e00c110",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "card_number": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    }
}